Update JwtJwksPrincipalProvider renaming userPermissionsClaim to rolesClaim after internal review.
Update JWTPrincipal to move code that was duplicated between Jwks and Shared Secret.

Update JwtSharedSecretPrincipalProvider to remove duplicate code.

Update pom.xml to reflect upcoming 2.1.8 version.
