+++ pom.xml
@@ -16,9 +16,9 @@
        release branch is labeled <major>.<minor>.<patch>-izgw-core-SNAPSHOT 
        main branch is labeled <major>.<minor>.<patch>-izgw-core-RELEASE 
   -->
-  <version>2.1.7-izgw-core-austin-SNAPSHOT</version>
+  <version>2.1.8-izgw-core-SNAPSHOT</version>
   <packaging>jar</packaging>
-  <name>IZ Gateway Core 2.1.7</name>
+  <name>IZ Gateway Core 2.1.8 </name>
   <description>IZ Gateway Core contains the core code for the IZ Gateway Hub and Transformation services</description>
   <distributionManagement>
     <repository>
